refactor tf_lstm.py
