+++ sips/lines/espn/espn_api.py
@@ -1,7 +1,4 @@
-import requests as r
-
 import sips.h.grab as g
-from sips.macros import macros as m
 from sips.lines import collate
 
 espn = "http://site.api.espn.com/apis/site/v2/sports/"
