+++ sips/macros/tfm.py
@@ -1,0 +1,5 @@
+from sips.macros import macros as m
+
+
+WRITE_TO = m.PROJ_DIR + "ml/logs/"
+READ_FROM = m.PARENT_DIR + "data/lines/lines/"
